requirements.txt
Flask==2.3.2
Flask-Cors==4.0.0
PyJWT==2.8.0
bcrypt==4.1.2
pymongo==4.5.0
python-dotenv==1.0.0
schedule==1.2.1
